requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
psycopg2-binary==2.9.10
gunicorn==19.9.0
Pillow==9.1.0
Flask-Mail==0.9.1
geoip2==4.7.0
pycolors==0.1.2
rfc3339==6.2
pytelegrambotapi==4.28.0
